Add a form of subtyping.
Specifically, allow "extension" of (co)data types.
The subsumption relation is the transitive closure
of the inverse of this extension relation.

Usefulness relies on the presence of a sensible
semantics for defaults. With respect to codata
types, this isn't present in the current
implementation of Uroboro.

(Also, only tested for very simple examples.)